Sign conversion fixes
Summary: Some type mismatches found by @pgriess.

Test Plan: fbmake runtests

Reviewed By: subodh@fb.com

Subscribers: trunkagent, pgriess, kmdent, fma, shikong, subodh, bmatheny, ranjeeth

FB internal diff: D1722229

Signature: t1:1722229:1417807133:669a7403ffbdb4292dbeea9b53cfd2a0634f127c

+++ proxygen/lib/http/HTTPCommonHeaders.template.h
@@ -42,7 +42,7 @@ enum HTTPHeaderCode : uint8_t {
 class HTTPCommonHeaders {
  public:
   // Perfect hash function to match common HTTP header names
-  static HTTPHeaderCode hash(const char* name, unsigned len);
+  static HTTPHeaderCode hash(const char* name, size_t len);
 
   inline static HTTPHeaderCode hash(const std::string& name) {
     return hash(name.data(), name.length());
